# src/integration-test/resources/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: api-installedbase
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "3.1"
servers:
  - url: http://127.0.0.1:8080
    description: Generated server url
tags:
  - name: Installed base
    description: Facility delegation operations
  - name: Installed base
    description: Installed base operations
paths:
  /{municipalityId}/delegates/{id}/{owner}:
    put:
      tags:
        - Installed base
      summary: Update a facility delegation
      description: "Update facilities, delegatedTo, and/or businessEngagementOrgId\
        \ on an existing facility delegation for a specific owner.To change the owner\
        \ or status of a delegation, use the DELETE and POST operations instead."
      operationId: putFacilityDelegation
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: id
          in: path
          description: Id of the delegation
          required: true
          schema:
            type: string
          example: 81471222-5798-11e9-ae24-57fa13b361e1
        - name: owner
          in: path
          description: Owner of the delegation
          required: true
          schema:
            type: string
          example: 81471222-5798-11e9-ae24-57fa13b361e1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFacilityDelegation"
        required: true
      responses:
        "204":
          description: No content
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/delegates:
    get:
      tags:
        - Installed base
      summary: Get delegation by owner and/or delegate
      operationId: getFacilityDelegations
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: owner
          in: query
          description: Owner of the delegation
          required: false
          schema:
            type: string
          example: 81471222-5798-11e9-ae24-57fa13b361e1
        - name: delegatedTo
          in: query
          description: The delegate
          required: false
          schema:
            type: string
          example: 81471222-5798-11e9-ae24-57fa13b361e2
        - name: status
          in: query
          description: "Status of the delegation, will show all delegation statuses\
          \ if not provided"
          required: false
          schema:
            type: string
          example: ACTIVE
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FacilityDelegationResponse"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
        - Installed base
      summary: Delegate facilities
      operationId: createFacilityDelegation
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFacilityDelegation"
        required: true
      responses:
        "201":
          description: Created
          headers:
            Location:
              style: simple
              schema:
                type: string
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/installedbase/{organizationNumber}:
    get:
      tags:
        - Installed base
      summary: Get installed base at company matching organization number for customer
        matching provided party-ID
      operationId: getInstalledBase
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: organizationNumber
          in: path
          description: Organization number
          required: true
          schema:
            type: string
          example: 5565112233
        - name: partyId
          in: query
          description: Party-ID
          required: true
          schema:
            type: array
            items:
              type: string
          example: 81471222-5798-11e9-ae24-57fa13b361e1
        - name: modifiedFrom
          in: query
          description: Optional date for filtering on installed bases modified at provided
            date or later
          required: false
          schema:
            type: string
            format: date
          example: 2023-06-01
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstalledBaseResponse"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/delegates/{id}:
    get:
      tags:
        - Installed base
      summary: Get a specific facility delegation
      operationId: getFacilityDelegationById
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: id
          in: path
          description: Id of the delegation
          required: true
          schema:
            type: string
          example: 81471222-5798-11e9-ae24-57fa13b361e1
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacilityDelegationResponse"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api-docs:
    get:
      tags:
        - API
      summary: OpenAPI
      operationId: getApiDocs
      responses:
        "200":
          description: OK
          content:
            application/yaml:
              schema:
                type: string
      x-auth-type: None
      x-throttling-tier: Unlimited
      x-wso2-mutual-ssl: Optional
components:
  schemas:
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
    StatusType:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        reasonPhrase:
          type: string
    UpdateFacilityDelegation:
      type: object
      properties:
        facilities:
          type: array
          items:
            minLength: 1
            type: string
        businessEngagementOrgId:
          type: string
          description: Organization number of the company owning the facility
          example: "5591628136"
        delegatedTo:
          type: string
          description: Party ID of the delegate
          example: 81471222-5798-11e9-ae24-57fa13b361e2
      description: UpdateFacilityDelegation model
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/StatusType"
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        message:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ThrowableProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    CreateFacilityDelegation:
      required:
        - delegatedTo
        - facilities
        - owner
      type: object
      properties:
        facilities:
          minItems: 1
          type: array
          items:
            minLength: 1
            type: string
        businessEngagementOrgId:
          type: string
          description: Organization number of the company owning the facility
          example: "5591628136"
        delegatedTo:
          type: string
          description: Party ID of the delegate
          example: 81471222-5798-11e9-ae24-57fa13b361e2
        owner:
          type: string
          description: Party ID of the delegation owner
          example: 81471222-5798-11e9-ae24-57fa13b361e1
      description: CreateFacilityDelegation model
    InstalledBaseCustomer:
      type: object
      properties:
        customerNumber:
          type: string
          description: Customer number
          readOnly: true
          example: "10007"
        partyId:
          type: string
          description: Party-ID
          readOnly: true
          example: cf9892ad-69d5-420f-ae98-9631dd1664fe
        items:
          type: array
          readOnly: true
          items:
            $ref: "#/components/schemas/InstalledBaseItem"
      description: Installed base owner model
      readOnly: true
    InstalledBaseItem:
      type: object
      properties:
        type:
          type: string
          description: Type
          readOnly: true
          example: Fjärrvärme
        facilityId:
          type: string
          description: Facility id
          readOnly: true
          example: "735999109270751042"
        placementId:
          type: integer
          description: Placement id
          format: int32
          readOnly: true
          example: 5263
        facilityCommitmentStartDate:
          type: string
          description: Facility commitment start date
          format: date
          readOnly: true
          example: 2020-04-01
        facilityCommitmentEndDate:
          type: string
          description: Facility commitment end date
          format: date
          readOnly: true
          example: 2020-09-30
        lastModifiedDate:
          type: string
          description: Last date for modification of item (or null if no modification
            has been done)
          format: date
          readOnly: true
          example: 2020-06-01
        address:
          $ref: "#/components/schemas/InstalledBaseItemAddress"
        metaData:
          maxItems: 1000
          type: array
          readOnly: true
          items:
            $ref: "#/components/schemas/InstalledBaseItemMetaData"
      description: Installed base item model
      readOnly: true
    InstalledBaseItemAddress:
      type: object
      properties:
        propertyDesignation:
          type: string
          description: Property designation
          readOnly: true
          example: Södermalm 1:27
        careOf:
          type: string
          description: Care of address
          readOnly: true
          example: Agatha Malm
        street:
          type: string
          description: Street
          readOnly: true
          example: Storgatan 9
        postalCode:
          type: string
          description: Postal code
          readOnly: true
          example: "85230"
        city:
          type: string
          description: City
          readOnly: true
          example: Sundsvall
      description: Installed base item address model
      readOnly: true
    InstalledBaseItemMetaData:
      type: object
      properties:
        key:
          type: string
          description: Key
          readOnly: true
          example: netarea
        value:
          type: string
          description: Value
          readOnly: true
          example: Sundsvall tätort
        type:
          type: string
          description: Type
          readOnly: true
          example: location
        displayName:
          type: string
          description: Displayname
          readOnly: true
          example: Nätområde
      description: Installed base item metadata model
      readOnly: true
    InstalledBaseResponse:
      type: object
      properties:
        installedBaseCustomers:
          type: array
          readOnly: true
          items:
            $ref: "#/components/schemas/InstalledBaseCustomer"
      description: Installed base response model
    FacilityDelegationResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the delegation
          example: 12345678-1234-1234-1234-123456789012
        facilities:
          type: array
          items:
            type: string
        businessEngagementOrgId:
          type: string
          description: Organization number of the company owning the facility
          example: "5591628136"
        delegatedTo:
          type: string
          description: Party ID of the delegate
          example: 81471222-5798-11e9-ae24-57fa13b361e2
        owner:
          type: string
          description: Party ID of the delegation owner
          example: 81471222-5798-11e9-ae24-57fa13b361e1
        status:
          type: string
          description: Status of the delegation
          example: ACTIVE
        municipalityId:
          type: string
          description: Municipality ID of the delegation
          example: "1234"
        created:
          type: string
          description: When the delegation was created
          format: date-time
        updated:
          type: string
          description: When the delegation was last updated
          format: date-time
        deleted:
          type: string
          description: When the delegation was deleted
          format: date-time
      description: FacilityDelegation response model
  securitySchemes: {}